refactor(http): extract HTTP/2 server logic to SocketHTTPServer-h2.c
Moves HTTP/2 stream handling to a dedicated module:
- Stream creation/lifecycle (server_http2_stream_get_or_create)
- Request building from HPACK headers (server_http2_build_request)
- Response sending (streaming and non-streaming)
- Flow control and window updates
- Stream callbacks and event handling
- HTTP/2 connection setup (server_http2_enable)

Main file reduced from 3768 to 2995 lines (-773 lines).
New h2 module: 806 lines.

Part of #1641
